# Simplify tree building code.