--- a/src/tree/updater_basemaker-inl.h
+++ b/src/tree/updater_basemaker-inl.h
@@ -333,28 +333,28 @@ class BaseMaker: public TreeUpdater {
     const MetaInfo &info = fmat.Info();
     thread_temp.resize(omp_get_max_threads());
     p_node_stats->resize(tree.param.num_nodes);
-    #pragma omp parallel
+#pragma omp parallel
     {
       const int tid = omp_get_thread_num();
-      thread_temp[tid].resize(tree.param.num_nodes, TStats(param_));
+      thread_temp[tid].resize(tree.param.num_nodes, TStats());
       for (unsigned int nid : qexpand_) {
-        thread_temp[tid][nid].Clear();
+        thread_temp[tid][nid] = TStats();
       }
     }
     // setup position
     const auto ndata = static_cast<bst_omp_uint>(fmat.Info().num_row_);
-    #pragma omp parallel for schedule(static)
+#pragma omp parallel for schedule(static)
     for (bst_omp_uint ridx = 0; ridx < ndata; ++ridx) {
       const int nid = position_[ridx];
       const int tid = omp_get_thread_num();
       if (nid >= 0) {
-        thread_temp[tid][nid].Add(gpair, info, ridx);
+        thread_temp[tid][nid].Add(gpair[ridx]);
       }
     }
     // sum the per thread statistics together
     for (int nid : qexpand_) {
       TStats &s = (*p_node_stats)[nid];
-      s.Clear();
+      s = TStats();
       for (size_t tid = 0; tid < thread_temp.size(); ++tid) {
         s.Add(thread_temp[tid][nid]);
       }