… src/tree/updater_colmaker.cc
… src/tree/updater_gpu_common.cuh
… src/tree/updater_histmaker.cc
… src/tree/updater_quantile_hist.cc
… src/tree/updater_quantile_hist.h
… src/tree/updater_refresh.cc
… src/tree/updater_skmaker.cc
… tests/cpp/tree/test_quantile_hist.cc