Remove:
- GHistRow
- GHistEntry
- GHistIndexRow
- kSimpleStats
- CheckInfo
- SetLeafVec

Moved LossChangeMissing output common since they are not common..
Addressed some compiler warnings, the build message will be cleaner.


There are suggestions from @hcho3 and @RAMitchell we reduce the stored data in algorithm class #3803, and I also want to structure these algorithms in a nicer way, so this would be a multi-stage refactor.

At first I will try to use Span class, if the bound check in it hurts performance too much I will switch to a similar class without bound check.  Along the way I also want to create unittests, which might be a little time consuming.