@trivialfis 2018-11-09
The things I really want to remove are `ValueConstrains` class and all `CalcGain` `CalcWeight` it uses.  But until we figure out how to get split evaluator working on GPU or find a better replacement, the duplication seems lingering.