@trivialfis 2018-11-12
It turns out, span does hurt the performace, following is the result from a simple benchmark.

Using dense dataset Epsilon, with following parameters

- nthread: 1
- tree_method: hist
- objective: binary:logistic

- training-data/full-dataset ratio set to 0.8 .

- CMake flags: -DCMAKE_BUILD_TYPE=RelWithDebInfo -DUSE_OPENMP=ON
- Platform: Ubuntu 18.04 x86-64
- CPU: AMD Ryzen 7 2700 Eight-Core Processor

# With first 5000 lines
Test error converge to 0 around 20 iterations.
## With span
78.37750792503357 sec
78.31656765937805 sec
## Without span
68.44277930259705
69.19314885139465

# With first 50000 lines
## With span
231.58369183540344 sec
237.2560956478119 sec
## Without span
218.65958523750305 sec
215.93592476844788 sec