@hcho3 2018-11-18
@trivialfis My work got really busy for the last week. The sprint will be over in a few days, then I should be able to add unit tests for the quantile hist updater. Can we merge this after that?