@hcho3 2019-01-04
I've re-based this PR against the latest master. Let's come back to it after #3957 is merged.