@trivialfis 2019-01-04
@hcho3 No problem, I looked into that PR, should be easy to rebase.