@trivialfis 2019-01-10
@hcho3 Ready for review.