@trivialfis 2019-01-19
@hcho3 Can you review this?  Since it modifys tree code so any other future PRs can have conflicts with this one.